backend/requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
fireworks-ai==0.15.6
astroid==3.3.5
urllib3==2.0.4